refactor: component 폴더를 파일로 분리 - SearchInput/index.tsx 형식이였던 컴포넌트를 SearchInput.tsx 형식으로 변경
