+++ src/components/Advertisement.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const index = () => {
+const Advertisement = () => {
     return <Container>Ad Section</Container>;
 };
 
@@ -12,4 +12,4 @@ const Container = styled.div`
     opacity: 0.1;
 `;
 
-export default index;
+export default Advertisement;
