+++ src/components/Pagination.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
-import { ReactComponent as ArrowLeft } from "../../assets/icons/arrow-left.svg";
-import { ReactComponent as ArrowRight } from "../../assets/icons/arrow-right.svg";
-import usePagination from "../../hooks/usePagination";
+import { ReactComponent as ArrowLeft } from "../assets/icons/arrow-left.svg";
+import { ReactComponent as ArrowRight } from "../assets/icons/arrow-right.svg";
+import usePagination from "../hooks/usePagination";
 import { useSearchParams } from "react-router-dom";
 
 type PaginationPropsType = {
@@ -15,7 +15,7 @@ type PageButtonPropsType = {
     activated?: boolean;
 };
 
-const index = ({ total, limit, maxPage }: PaginationPropsType) => {
+const Pagination = ({ total, limit, maxPage }: PaginationPropsType) => {
     const [params, setParams] = useSearchParams();
     const { prevPage, nextPage, pageRange, totalPages } = usePagination(
         total,
@@ -93,4 +93,4 @@ const PageButton = styled.button<PageButtonPropsType>`
         `background-color: ${theme.pageBtnActive}; border-radius: 10px; color: white;`};
 `;
 
-export default index;
+export default Pagination;
